requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Pillow==10.0.0
bcrypt==4.0.1
PyJWT==2.8.0
markdown==3.4.4
Werkzeug==2.3.7
email-validator==2.0.0.post2
PyMySQL==1.1.0
flask-mysqldb==2.0.0
mysqlclient==2.2.0
itsdangerous==2.1.2
gunicorn==21.2.0
